refactor util method
